+++ src/firebase/proposals.js
@@ -1,7 +1,5 @@
 import firebase from 'firebase/app'
 import omit from 'lodash/omit'
-import toLower from 'lodash/toLower'
-import deburr from 'lodash/deburr'
 
 /**
  * Return the proposal with the given id
@@ -25,12 +23,7 @@ export const fetchEventProposals = async (
   eventId,
   uid,
   {
-    categories,
-    formats,
-    state,
-    sortOrder,
-    ratings,
-    search,
+    categories, formats, state, sortOrder, ratings,
   } = {},
 ) => {
   let query = firebase
@@ -65,12 +58,6 @@ export const fetchEventProposals = async (
   const result = await query.get()
   let proposals = result.docs.map(ref => ({ id: ref.id, ...ref.data() }))
 
-  // add search by title (client filter)
-  if (search) {
-    const searchQuery = deburr(toLower(search))
-    proposals = proposals.filter(proposal => deburr(toLower(proposal.title)).includes(searchQuery))
-  }
-
   // add ratings filter (client filter)
   if (ratings === 'rated') {
     proposals = proposals.filter(proposal => proposal.usersRatings && !!proposal.usersRatings[uid])
@@ -95,12 +82,40 @@ export const updateProposal = (eventId, proposal, options = {}) => {
     .update(updated)
 }
 
-export const updateRating = (eventId, talkId, uid, ratingUpdated, rated) => {
+export const updateRating = (eventId, proposalId, uid, ratingUpdated, rated) => {
   firebase
     .firestore()
     .collection('events')
     .doc(eventId)
     .collection('proposals')
-    .doc(talkId)
+    .doc(proposalId)
     .update({ ...ratingUpdated, [`usersRatings.${uid}`]: rated })
+}
+
+export const fetchOrganizersThread = async (eventId, proposalId) => {
+  const result = await firebase
+    .firestore()
+    .collection('events')
+    .doc(eventId)
+    .collection('proposals')
+    .doc(proposalId)
+    .collection('organizersThread')
+    .orderBy('date', 'asc')
+    .get()
+  return result.docs.map(ref => ref.data())
+}
+
+export const addOrganizersThreadMessage = async (eventId, proposalId, uid, message) => {
+  await firebase
+    .firestore()
+    .collection('events')
+    .doc(eventId)
+    .collection('proposals')
+    .doc(proposalId)
+    .collection('organizersThread')
+    .add({
+      uid,
+      message,
+      date: firebase.firestore.FieldValue.serverTimestamp(),
+    })
 }