+++ src/store/drivers/index.js
@@ -1,7 +1,0 @@
-import router from './router'
-import form from './redux-form'
-
-export default {
-  router,
-  form,
-}